cooldown reset for bumpers
