+++ cogs/stats.py
@@ -6,34 +6,12 @@ from discord import app_commands
 from discord import ui
 from typing import Optional
 
-from classes.checks import is_premium_server, is_premium
+from classes.cooldown import default_cooldown, hard_cooldown
 from classes import checks
 from classes import db
 from config import *
 
 logger = logging.getLogger("discord")
-
-
-async def default_cooldown(
-    interaction: discord.Interaction,
-) -> Optional[app_commands.Cooldown]:
-    assert interaction.guild is not None
-    if await is_premium(interaction.user.id) != "None" or await is_premium_server(
-        interaction.guild
-    ):
-        return None
-    return app_commands.Cooldown(1, 3.0)
-
-
-async def hard_cooldown(
-    interaction: discord.Interaction,
-) -> Optional[app_commands.Cooldown]:
-    assert interaction.guild is not None
-    if await is_premium(interaction.user.id) != "None" or await is_premium_server(
-        interaction.guild
-    ):
-        return app_commands.Cooldown(1, 2.0)
-    return app_commands.Cooldown(1, 10.0)
 
 
 class Stats(commands.Cog):
